updating viewer v2 as main viewer
