+++ lib/pdf_viewer.dart
@@ -4,4 +4,3 @@ export 'src/document.dart' show PDFDocument;
 export 'src/page.dart' show PDFPage;
 export 'src/viewer.dart' show PDFViewer;
 export 'src/tooltip.dart' show PDFViewerTooltip;
-export 'src/viewer_v2.dart' show PDFViewerV2;
